backend/requirements.txt
# FastAPI & Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
greenlet==3.0.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Mercado Pago
mercadopago==2.2.1

# Utils
python-slugify==8.0.2
pillow==10.2.0
aiofiles==23.2.1
httpx==0.26.0

# Email
aiosmtplib==3.0.1
jinja2==3.1.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
aiosqlite==0.19.0
httpx==0.26.0

# Dev
python-dotenv==1.0.1
